pyproject.toml
[tool.poetry]
name = "fast_zero"
version = "0.1.0"
description = ""
authors = ["jorge.machado-Kraken <jorge.vmachado@hotmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = {extras = ["standard"], version = "^0.128.1"}
sqlalchemy = "^2.0.46"
pydantic-settings = "^2.12.0"
alembic = "^1.18.3"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
pyjwt = "^2.11.0"
tzdata = "^2025.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.0,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]
[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -x --cov=fast_zero -vv'
post_test = 'coverage html'